Correct make distcheck and most gcc/sparse warnings.
  Remaining warnings are due to bit/shift operations mixed with
addition/multiplication/etc. "1<<2-1" should be read as
"1 << (2 - 1)", and not "(1 << 2) - 1". The first 2 warnings
appear to be for a confusing case, as a few lines bellow
it is adding parenthesis to the shift operation, but should be
being parsed correctly (not added parenthesis to avoid the
adding them wrong, and "hiding" the warning...)